+++ components/Days.vue
@@ -72,8 +72,7 @@ export default {
   justify-content: space-around;
   align-items: center;
   padding: 35px 60px;
-  border: var(--border-color) solid;
-  border-width: 1px;
+  border: 1px var(--border-color) solid;
 }
 .day-box {
   display: flex;
