+++ components/Goals.vue
@@ -74,8 +74,7 @@ export default {
   background-color: var(--background-card);
   margin: 10px 0px;
   padding: 25px 30px;
-  border: var(--border-color) solid;
-  border-width: 1px;
+  border: 1px var(--border-color) solid;
 }
 .goals-item:hover {
   transform: scale(1.005);
