+++ helpers/document.js
@@ -1,0 +1,3 @@
+export function addClassToDocument(className) {
+  document.documentElement.className = className
+}
